refactored the files file into smaller chunks
